Factorizacion + sistema de HP enemigo
Se ha refactorizado el código que controla a los enemigos de manera que existe una clase padre 'Enemigo', de las que heredan los distintos tipos de enemigo que hay (melee y ranged). Esto facilita la implementación de  otras mecánicas y de nuevos enemigos en el futuro.

Además, se ha incorporado el sistema de HP enemigo, junto al comportamiento que estos tienen al recibir daños. Falta implementarlo visualmente, para lo que es necesario el sistema de fijado.

+++ Assets/Characters/Enemies/CYAN ENEMY/Behaviour Scripts/CYAN_Attack_Behaviour.cs
@@ -10,7 +10,7 @@ public class CYAN_Attack_Behaviour : StateMachineBehaviour
 
     [SerializeField] float attackImpulse;
     [SerializeField] float attackSpeed;
-    [SerializeField] float attackDamage;
+    [SerializeField] int attackDamage;
 
     // OnStateEnter is called when a transition starts and the state machine starts to evaluate this state
     override public void OnStateEnter(Animator animator, AnimatorStateInfo stateInfo, int layerIndex)
